% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_server.R
\name{main_server}
\alias{main_server}
\title{Main Server}
\usage{
main_server()
}
\description{
This function returns the main server function for this dashboard.
}
\examples{
\dontrun{
shiny::shinyApp(main_ui(), main_server())
}
}
