% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality_codes.R
\name{read_municipality}
\alias{read_municipality}
\title{Municipality codes and names}
\usage{
read_municipality()
}
\value{
a dataframe with every municipality code,name and province
}
\description{
Reads the identification data for municipalities
}
\examples{
df = mun_code_to_name()
}
