% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Start Dashboard}
\usage{
run(woondata = file.path(getwd(), "cache", "WoON2018_e_1.0.sav"))
}
\arguments{
\item{woondata}{The path to the woon data file}
}
\description{
This function starts the dashboard.
}
\examples{
\dontrun{
run()
}
}
